# api.yaml
openapi: 3.0.0
info:
  title: Plaid SDK Assignment
  version: 1.0.0
servers:
  - url: "http://localhost:3000"
paths:
  /institutions/search:
    post:
      parameters:
        - in: header
          name: PLAID-CLIENT-ID
          required: false # Set to false if not required
          schema:
            type: string
        - in: header
          name: PLAID-SECRET
          required: false # Set to false if not required
          schema:
            type: string
      summary: list institutions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                query:
                  type: string
                  description: "The search query. Institutions with names matching the query are returned"
                country_codes:
                  type: array
                  items:
                    type: string
                    enum:
                      - US
                      - GB
                      - ES
                      - NL
                      - FR
                      - IE
                      - CA
                      - DE
                      - IT
                      - PL
                      - DK
                      - NO
                      - SE
                      - EE
                      - LT
                      - LV
                      - PT
                      - BE
                  description: "Specify which country or countries to include institutions from"
                  example: ["US", "GB"]
                products:
                  type: array
                  items:
                    type: string
                    enum:
                      - assets
                      - auth
                      - balance
                      - employment
                      - identity
                      - income_verification
                      - investments
                      - liabilities
                      - identity_verification
                      - payment_initiation
                      - standing_orders
                      - transactions
                      - transfer
                  description: "Filter the Institutions based on whether they support all products listed in products"
                  example: ["transactions", "auth"]
                options:
                  type: object
                  description: "An optional object to filter /institutions/search results"
                  properties:
                    oauth:
                      type: boolean
                      description: "Limit results to institutions with or without OAuth login flows."
                    include_optional_metadata:
                      type: boolean
                      description: "When true, return the institution's homepage URL, logo, and primary brand color."
                    include_auth_metadata:
                      type: boolean
                      description: "When true, returns metadata related to the Auth product indicating which auth methods are supported."
                      default: false
                    include_payment_initiation_metadata:
                      type: boolean
                      description: "When true, returns metadata related to the Payment Initiation product indicating which payment configurations are supported."
                      default: false
                    payment_initiation:
                      type: object
                      description: "Additional options to filter institutions by various Payment Initiation configurations."
                      properties:
                        payment_id:
                          type: string
                          description: "A unique ID identifying the payment."
                        consent_id:
                          type: string
                          description: "A unique ID identifying the payment consent."
              required:
                - query
                - country_codes
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  institutions:
                    type: array
                    items:
                      type: object
                      properties:
                        country_codes:
                          type: array
                          items:
                            type: string
                        dtc_numbers:
                          type: array
                          items:
                            type: string
                        institution_id:
                          type: string
                        name:
                          type: string
                        oauth:
                          type: boolean
                        products:
                          type: array
                          items:
                            type: string
                        routing_numbers:
                          type: array
                          items:
                            type: string
                  request_id:
                    type: string
        "400":
          description: Invalid request
        "500":
          description: Server error
      security:
        - clientId: []
        - secret: []
components:
  securitySchemes:
    clientId:
      type: apiKey
      in: header
      name: PLAID-CLIENT-ID
    secret:
      type: apiKey
      in: header
      name: PLAID-SECRET